pyproject.toml
[project]
name = "qvapaysdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.2",
]
